Kickstarter platform Predictive Analytics project
Kickstarter is an American public-benefit corporation, that maintains a global crowdfunding platform focused on creativity and merchandising. Our objective is to explore data and present the most interesting findings. To understand this data we conducted descriptive analysis using Python and SAS along with visualizations using Tableau Dashboards. While performing descriptive analysis, we checked the independence of the variables using statistical tests and noted the correlation present among variables. We have also tried scaling and normalizing few data columns in order to gain better understanding of the data.